requirements.txt
pyautogui
pillow
tk
opencv-python